pyproject.toml
[project]
name = "deepagent"
version = "0.1.0"
description = "AI-powered planning assistant that transforms goals into actionable plans"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepagents>=0.2.5",
    "langchain-openai>=1.0.2",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.12",
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["api", "main", "prompts"]
